requirements.txt
groq
python-dotenv
Pillow
